Rename `com.fsck.k9.preferences.migrations` to `….migration`
