+++ app/storage/src/main/java/com/fsck/k9/preferences/migration/StorageMigrationTo6.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.preferences.migrations
+package com.fsck.k9.preferences.migration
 
 import android.database.sqlite.SQLiteDatabase
