+++ app/storage/src/main/java/com/fsck/k9/preferences/migration/StorageMigrations.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.preferences.migrations
+package com.fsck.k9.preferences.migration
 
 import android.database.sqlite.SQLiteDatabase
