+++ app/storage/src/main/java/com/fsck/k9/preferences/migration/migration12/Pop3StoreUriDecoder.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.preferences.migrations.migration12;
+package com.fsck.k9.preferences.migration.migration12;
 
 
 import java.net.URI;
