+++ app/storage/src/main/java/com/fsck/k9/preferences/migration/migration12/WebDavStoreUriDecoder.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.preferences.migrations.migration12;
+package com.fsck.k9.preferences.migration.migration12;
 
 
 import com.fsck.k9.mail.AuthType;
