+++ app/storage/src/test/java/com/fsck/k9/preferences/migration/StorageMigrationTo22Test.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.preferences.migrations
+package com.fsck.k9.preferences.migration
 
 import assertk.all
 import assertk.assertThat
